% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfun.R
\name{change2}
\alias{change2}
\title{A function that separates the target function and the constraints}
\usage{
change2(mydata)
}
\arguments{
\item{mydata}{A matrix}
}
\value{
the separated matrix from \code{\link{return}}
}
\description{
A function that separates the target function and the constraints
}
\examples{
 \dontrun{
# No need to run this
}
x <- cbind(runif(20,1,5),runif(20,1,5),runif(20,1,5))
change2(x)
}
