% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfun.R
\name{conssym}
\alias{conssym}
\title{The function to derive the symmetric constraints}
\usage{
conssym(n)
}
\arguments{
\item{n}{}
}
\value{
the symmetry constraints from \code{\link{return}}
}
\description{
The function to derive the symmetric constraints
}
\examples{
\dontrun{
# No need to run this
}
conssym(3)
}
