% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfun.R
\name{crossderiv1}
\alias{crossderiv1}
\title{calculate the 1st order derivative of cross products}
\usage{
crossderiv1(x)
}
\arguments{
\item{x}{}
}
\value{
the 1st order production terms from \code{\link{return}}
}
\description{
calculate the 1st order derivative of cross products
}
\examples{
\dontrun{
# No need to run this
}
x <- data.frame(runif(20,1,5),runif(20,1,5),runif(20,1,5))
crossderiv1(x)
}
