% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfun.R
\name{crossderiv2}
\alias{crossderiv2}
\title{calculate the 1st order derivative of cross products of 2 matrices or dataframes}
\usage{
crossderiv2(x, y)
}
\arguments{
\item{x}{the first matrix}

\item{y}{the second matrix}
}
\value{
the 1st order production terms from \code{\link{return}}
}
\description{
calculate the 1st order derivative of cross products of 2 matrices or dataframes
}
\examples{
\dontrun{
# No need to run this
}
x <- data.frame(runif(20,1,5),runif(20,1,5),runif(20,1,5))
y <- data.frame(runif(20,1,5),runif(20,1,5))
crossderiv1(x)
}
