% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfun.R
\name{rowprod}
\alias{rowprod}
\title{a function that calculate the column cross production terms for a matrix or dataframe}
\usage{
rowprod(mydata)
}
\arguments{
\item{mydata}{}
}
\value{
the column cross production terms from \code{\link{return}}
}
\description{
a function that calculate the column cross production terms for a matrix or dataframe
}
\examples{
x <- data.frame(runif(20,1,5),runif(20,1,5),runif(20,1,5))
rowprod(x)
}
