% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfun.R
\name{rowprod2}
\alias{rowprod2}
\title{a function that calculate the column cross production terms for 2 matrices or dataframes}
\usage{
rowprod2(mydata1, mydata2)
}
\arguments{
\item{mydata1}{matrix or dataframe 1}

\item{mydata2}{matrix or dataframe 2}
}
\value{
the column cross production terms from \code{\link{return}}
}
\description{
a function that calculate the column cross production terms for 2 matrices or dataframes
}
\examples{
x <- data.frame(runif(20,1,5),runif(20,1,5),runif(20,1,5))
y <- data.frame(runif(20,1,5),runif(20,1,5),runif(20,1,5))
rowprod2(x,y)
}
